Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

SolverTest.bound_constrained_nlp

SolverTest.bound_constrained_nls

SolverTest.equality_constrained_nlp

SolverTest.equality_constrained_nls

SolverTest.kkt_checker

SolverTest.multiprecision_nlp

SolverTest.multiprecision_nls

SolverTest.unconstrained_nlp

SolverTest.unconstrained_nls

​

SolverTest.bound_constrained_nlp — Method

bound_constrained_nlp(solver; problem_set = bound_constrained_nlp_set(), atol = 1e-6, rtol = 1e-6, kwargs...)

Test the solver on bound-constrained problems. If rtol is non-zero, the relative error uses the gradient at the initial guess. Additional keyword arguments are passed to the solver.

SolverTest.bound_constrained_nls — Method

bound_constrained_nls(solver; problem_set = bound_constrained_nls_set(), atol = 1e-6, rtol = 1e-6, kwargs...)

Test the solver on bound-constrained nonlinear least-squares problems. If rtol is non-zero, the relative error uses the gradient at the initial guess. Additional keyword arguments are passed to the solver.

SolverTest.equality_constrained_nlp — Method

equality_constrained_nlp(solver; problem_set = equality_constrained_nlp_set(), atol = 1e-6, rtol = 1e-6, kwargs...)

Test the solver on equality-constrained problems. If rtol is non-zero, the relative error uses the gradient at the initial guess. Additional keyword arguments are passed to the solver.

SolverTest.equality_constrained_nls — Method

equality_constrained_nls(solver; problem_set = equality_constrained_nls_set(), atol = 1e-6, rtol = 1e-6, kwargs...)

Test the solver on equality-constrained problems. If rtol is non-zero, the relative error uses the gradient at the initial guess. Additional keyword arguments are passed to the solver.

SolverTest.kkt_checker — Method

kkt_checker(nlp, sol; kwargs...)

Given an NLPModels nlp and a vector sol, it returns the KKT residual of an optimization problem as a tuple (primal, dual). In particular, it uses ripqp to solve the following quadratic optimization problem with linear constraints

min_{d} ∇f(sol)ᵀd +  ½∥d∥²
        lvar ≤ sol + d ≤ uvar
        lcon ≤ c(sol) + ∇c(sol)d ≤ ucon

The solution of this problem is the gradient of the Lagrangian of the nlp at sol thanks to the ½ ‖d‖² term in the objective.

Keyword arguments are passed to RipQP.

SolverTest.multiprecision_nlp — Method

multiprecision_nlp(solver, problem_type; precisions=[Float16, …, BigFloat])

Test that solver solves a problem of type problem_type on various precisions. The problem_type can be

:unc

:bnd

:equ

:ineq

:eqnbnd

:gen

SolverTest.multiprecision_nls — Method

multiprecision_nls(solver, problem_type; precisions=[Float16, …, BigFloat])

Test that solver solves a problem of type problem_type on various precisions. The problem_type can be

:unc

:bnd

:equ

:ineq

:eqnbnd

:gen

SolverTest.unconstrained_nlp — Method

unconstrained_nlp(solver; problem_set = unconstrained_nlp_set(), atol = 1e-6, rtol = 1e-6, kwargs...)

Test the solver on unconstrained problems. If rtol is non-zero, the relative error uses the gradient at the initial guess. Additional keyword arguments are passed to the solver.

SolverTest.unconstrained_nls — Method

unconstrained_nls(solver; problem_set = unconstrained_nls_set(), atol = 1e-6, rtol = 1e-6, kwargs...)

Test the solver on unconstrained nonlinear least-squares problems. If rtol is non-zero, the relative error uses the gradient at the initial guess. Additional keyword arguments are passed to the solver.